tests/wpt_dom/tests/smoke/sync-fail.html

<!doctype html>
<meta charset="utf-8" />
<title>WPT DOM smoke: sync fail (intentional)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  async_test((t) => {
    t.step_func_done(() => {
      assert_true(false, "intentional failing assertion");
    })();
  }, "intentional failing assertion");
</script>
